+++ compose.yaml
@@ -40,6 +40,7 @@ services:
 
   block_committer:
     container_name: block-committer
+    restart: unless-stopped
     depends_on:
       fuel_node:
         condition: service_healthy
@@ -51,15 +52,17 @@ services:
     ports:
       - "8080:8080"
     environment:
-      - FUEL_GRAPHQL_ENDPOINT=http://fuel_node:4000
-      - ETHEREUM_RPC=ws://eth_node:8545
-      - POSTGRES_HOST=postgres
-      - POSTGRES_USER=username
-      - POSTGRES_PASS=password
-      - POSTGRES_DB=test
+      - COMMITTER__FUEL__GRAPHQL_ENDPOINT=http://fuel_node:4000
+      - COMMITTER__ETH__RPC=ws://eth_node:8545
+      - COMMITTER__APP__DB__HOST=postgres
     volumes:
       - ./configurations/development/config.toml:/app/config.toml
-    command: ["./fuel-block-committer", "--config", "/app/config.toml"]
+    command: ["/app/config.toml"]
+    healthcheck:
+      test: [ "CMD", "curl", "--fail", "http://0.0.0.0:8080/health" ]
+      retries: 50
+      interval: 2s
+      timeout: 30s
     
 
   postgres:
