+++ configurations/development/config.toml
@@ -1,21 +1,21 @@
 [eth]
 wallet_key = "0x9e56ccf010fa4073274b8177ccaad46fbaf286645310d03ac9bb6afa922a7c36"
-rpc = "${ETHEREUM_RPC}"
-chain_id = "anvil"
+chain_id = "hardhat"
 state_contract_address = "0xdAad669b06d79Cb48C8cfef789972436dBe6F24d"
 commit_interval = 3
-
-[fuel]
-gql_address = "${FUEL_GRAPHQL_ENDPOINT}"
+rpc = "ws://localhost:8545"
 
 [app]
 port = 8080
 host = "0.0.0.0"
 
+[fuel]
+graphql_endpoint = "http://localhost:4000"
+
 [app.db]
-host = "${POSTGRES_HOST}"
+host = "localhost"
 port = 5432
-username = "${POSTGRES_USER}"
-password = "${POSTGRES_PASS}"
-db = "${POSTGRES_DB}"
-connections = 5
+username = "username"
+password = "password"
+database = "test"
+max_connections = 5
