+++ deployment/Dockerfile
@@ -25,7 +25,4 @@ WORKDIR /root/
 COPY --from=builder /build/target/release/fuel-block-committer .
 COPY --from=builder /build/target/release/fuel-block-committer.d .
 
-# https://stackoverflow.com/a/44671685
-# https://stackoverflow.com/a/40454758
-# hadolint ignore=DL3025
-CMD ["./fuel-block-committer"]
+ENTRYPOINT ["./fuel-block-committer"]
