+++ src/main.rs
@@ -1,3 +1,4 @@
+#![deny(unused_crate_dependencies)]
 mod adapters;
 mod api;
 mod config;
@@ -7,14 +8,12 @@ mod setup;
 mod telemetry;
 
 use api::launch_api_server;
+use config::InternalConfig;
 use errors::Result;
 use prometheus::Registry;
-use setup::{
-    config::InternalConfig,
-    helpers::{
-        create_eth_adapter, setup_logger, setup_storage, shut_down, spawn_block_watcher,
-        spawn_eth_committer_and_listener, spawn_wallet_balance_tracker,
-    },
+use setup::helpers::{
+    create_eth_adapter, setup_logger, setup_storage, shut_down, spawn_block_watcher,
+    spawn_eth_committer_and_listener, spawn_wallet_balance_tracker,
 };
 use tokio_util::sync::CancellationToken;
 
@@ -23,10 +22,12 @@ async fn main() -> Result<()> {
     setup_logger();
 
     let config = config::parse()?;
-    let internal_config = InternalConfig::default();
-    let cancel_token = CancellationToken::new();
 
     let storage = setup_storage(&config).await?;
+    storage.migrate().await?;
+
+    let internal_config = InternalConfig::default();
+    let cancel_token = CancellationToken::new();
 
     let metrics_registry = Registry::default();
 
@@ -75,4 +76,10 @@ async fn main() -> Result<()> {
         listener_handle,
     )
     .await
+}
+
+#[cfg(test)]
+mod tests {
+    // used in the harness
+    use anyhow as _;
 }
