+++ src/setup.rs
@@ -1,2 +1,1 @@
-pub mod config;
 pub mod helpers;
